+++ packages/core/src/components/Canvas/Canvas.tsx
@@ -1,6 +1,6 @@
-import { observer } from 'mobx-react-lite'
 import * as React from 'react'
 import { Brush } from '~components/Brush'
+import { Cursor, CursorComponent } from '~components/Cursor'
 import { EraseLine } from '~components/EraseLine'
 import { Grid } from '~components/Grid'
 import { Overlay } from '~components/Overlay'
@@ -30,14 +30,14 @@ import type {
   TLUsers,
 } from '~types'
 
-interface CanvasProps<T extends TLShape, M extends Record<string, unknown>> {
+export interface CanvasProps<T extends TLShape, M extends Record<string, unknown>> {
   page: TLPage<T, TLBinding>
   pageState: TLPageState
   assets: TLAssets
   snapLines?: TLSnapLine[]
   eraseLine?: number[][]
   grid?: number
-  users?: TLUsers<T>
+  users?: TLUsers
   userId?: string
   hideBounds: boolean
   hideHandles: boolean
@@ -50,15 +50,16 @@ interface CanvasProps<T extends TLShape, M extends Record<string, unknown>> {
   showDashedBrush: boolean
   externalContainerRef?: React.RefObject<HTMLElement>
   performanceMode?: TLPerformanceMode
+  components?: {
+    Cursor?: CursorComponent
+  }
   meta?: M
   id?: string
   onBoundsChange: (bounds: TLBounds) => void
+  hideCursors?: boolean
 }
 
-export const Canvas = observer(function _Canvas<
-  T extends TLShape,
-  M extends Record<string, unknown>
->({
+function _Canvas<T extends TLShape, M extends Record<string, unknown>>({
   id,
   page,
   pageState,
@@ -68,9 +69,9 @@ export const Canvas = observer(function _Canvas<
   grid,
   users,
   userId,
+  components = {},
   meta,
   performanceMode,
-  externalContainerRef,
   showDashedBrush,
   hideHandles,
   hideBounds,
@@ -81,6 +82,7 @@ export const Canvas = observer(function _Canvas<
   hideRotateHandle,
   hideGrid,
   onBoundsChange,
+  hideCursors,
 }: CanvasProps<T, M>) {
   const rCanvas = React.useRef<HTMLDivElement>(null)
 
@@ -88,7 +90,7 @@ export const Canvas = observer(function _Canvas<
 
   rZoomRef.current = pageState.camera.zoom
 
-  useZoomEvents(rZoomRef, externalContainerRef || rCanvas)
+  useZoomEvents(rZoomRef, rCanvas)
 
   useResizeObserver(rCanvas, onBoundsChange)
 
@@ -132,7 +134,9 @@ export const Canvas = observer(function _Canvas<
           {pageState.brush && (
             <Brush brush={pageState.brush} dashed={showDashedBrush} zoom={pageState.camera.zoom} />
           )}
-          {users && <Users userId={userId} users={users} />}
+          {users && !hideCursors && (
+            <Users userId={userId} users={users} Cursor={components?.Cursor ?? Cursor} />
+          )}
         </div>
         <Overlay camera={pageState.camera}>
           {eraseLine && <EraseLine points={eraseLine} zoom={pageState.camera.zoom} />}
@@ -141,4 +145,6 @@ export const Canvas = observer(function _Canvas<
       </div>
     </div>
   )
-})
+}
+
+export const Canvas = React.memo(_Canvas)