+++ packages/core/src/hooks/useZoomEvents.ts
@@ -36,7 +36,7 @@ export function useZoomEvents<T extends HTMLElement>(
 
       const [x, y, z] = normalizeWheel(e)
 
-      // alt+scroll or ctrl+scroll = zoom
+      // alt+scroll or ctrl+scroll = zoom (when not clicking)
       if ((e.altKey || e.ctrlKey || e.metaKey) && e.buttons === 0) {
         const point = inputs.pointer?.point ?? [bounds.width / 2, bounds.height / 2]
         const delta = [...point, z * 0.618]
@@ -48,7 +48,7 @@ export function useZoomEvents<T extends HTMLElement>(
 
       // otherwise pan
       const delta = Vec.mul(
-        e.shiftKey && !Utils.isDarwin
+        e.shiftKey && !Utils.isDarwin()
           ? // shift+scroll = pan horizontally
             [y, 0]
           : // scroll = pan vertically (or in any direction on a trackpad)