+++ packages/tldraw/src/hooks/useKeyboardShortcuts.tsx
@@ -199,8 +199,8 @@ export function useKeyboardShortcuts(ref: React.RefObject<HTMLDivElement>) {
   useHotkeys(
     'ctrl+n,⌘+n',
     (e) => {
+      e.preventDefault()
       if (!canHandleEvent()) return
-
       onNewProject(e)
     },
     undefined,
@@ -307,7 +307,7 @@ export function useKeyboardShortcuts(ref: React.RefObject<HTMLDivElement>) {
   // Camera
 
   useHotkeys(
-    'ctrl+=,⌘+=,ctrl+num_subtract,⌘+num_subtract',
+    'ctrl+=,⌘+=,ctrl+num_add,⌘+num_add',
     (e) => {
       if (!canHandleEvent(true)) return
       app.zoomIn()
@@ -318,7 +318,7 @@ export function useKeyboardShortcuts(ref: React.RefObject<HTMLDivElement>) {
   )
 
   useHotkeys(
-    'ctrl+-,⌘+-,ctrl+num_add,⌘+num_add',
+    'ctrl+-,⌘+-,ctrl+num_subtract,⌘+num_subtract',
     (e) => {
       if (!canHandleEvent(true)) return