+++ packages/tldraw/src/state/shapes/TextUtil/TextUtil.tsx
@@ -8,6 +8,8 @@ import { TDShapeUtil } from '~state/shapes/TDShapeUtil'
 import {
   TextAreaUtils,
   defaultTextStyle,
+  getFontFace,
+  getFontSize,
   getFontStyle,
   getShapeStyle,
   getTextAlign,
@@ -204,8 +206,11 @@ export class TextUtil extends TDShapeUtil<T, E> {
         [isEditing]
       )
 
+      const rWasEditing = React.useRef(isEditing)
+
       React.useEffect(() => {
         if (isEditing) {
+          rWasEditing.current = true
           this.texts.set(shape.id, text)
           requestAnimationFrame(() => {
             rIsMounted.current = true
@@ -215,7 +220,8 @@ export class TextUtil extends TDShapeUtil<T, E> {
               elm.select()
             }
           })
-        } else {
+        } else if (rWasEditing.current) {
+          rWasEditing.current = false
           onShapeBlur?.()
         }
       }, [isEditing])
@@ -389,10 +395,23 @@ export class TextUtil extends TDShapeUtil<T, E> {
   getSvgElement = (shape: T, isDarkMode: boolean): SVGElement | void => {
     const bounds = this.getBounds(shape)
     const style = getShapeStyle(shape.style, isDarkMode)
-    const elm = getTextSvgElement(shape.text, shape.style, bounds)
-    elm.setAttribute('fill', style.stroke)
 
-    return elm
+    const fontSize = getFontSize(shape.style.size, shape.style.font) * (shape.style.scale ?? 1)
+    const fontFamily = getFontFace(shape.style.font).slice(1, -1)
+    const textAlign = shape.style.textAlign ?? AlignStyle.Start
+
+    const textElm = getTextSvgElement(
+      shape.text,
+      fontSize,
+      fontFamily,
+      textAlign,
+      bounds.width,
+      false
+    )
+
+    textElm.setAttribute('fill', style.stroke)
+
+    return textElm
   }
 }