+++ packages/tldraw/src/types.ts
@@ -176,7 +176,7 @@ export enum TDUserStatus {
 }
 
 // A TDUser, for multiplayer rooms
-export interface TDUser extends TLUser<TDShape> {
+export interface TDUser extends TLUser {
   activeShapes: TDShape[]
   status: TDUserStatus
   session?: boolean
@@ -460,7 +460,7 @@ export enum AlignStyle {
 export enum FontStyle {
   Script = 'script',
   Sans = 'sans',
-  Serif = 'erif',
+  Serif = 'serif',
   Mono = 'mono',
 }