+++ packages/tldraw/CHANGELOG.md
@@ -1,5 +1,149 @@
 # Changelog
 
+## 1.26.2
+
+### Patch Changes
+
+- Fix types in core.
+- Updated dependencies
+  - @tldraw/core@1.20.1
+
+## 1.26.1
+
+### Patch Changes
+
+- Fixes blurring text inputs in multiplayer.
+
+## 1.26.0
+
+### Minor Changes
+
+- Add metadata property to user.
+
+### Patch Changes
+
+- Updated dependencies
+  - @tldraw/core@1.20.0
+
+## 1.25.0
+
+### Minor Changes
+
+- - Adds `components` prop to Tldraw component (for custom Cursor, etc) @jamesbvaughan
+  - Adds Thai language @watchakorn-18k
+  - Fix event bug on `onRightPointCanvas`
+  - Fix bug with bad data in document with up-to-date version number
+  - Fix bug with arrow bindings
+  - Improves freehand line performance
+  - Improves performance of shape tree
+  - Improved .tldr file size (strip white space)
+
+### Patch Changes
+
+- Updated dependencies
+  - @tldraw/core@1.19.0
+  - @tldraw/intersect@1.8.0
+  - @tldraw/vec@1.8.0
+
+## 1.24.5
+
+### Patch Changes
+
+- Fix mouse events.
+- Updated dependencies
+  - @tldraw/core@1.18.4
+
+## 1.24.4
+
+### Patch Changes
+
+- Fix menu bug.
+
+## 1.24.3
+
+### Patch Changes
+
+- Small bump.
+- Updated dependencies
+  - @tldraw/core@1.18.3
+
+## 1.24.2
+
+### Patch Changes
+
+- Fix bug with scrolling.
+- Updated dependencies
+  - @tldraw/core@1.18.2
+
+## 1.24.1
+
+### Patch Changes
+
+- - Fix bug with mouse button state
+- Updated dependencies
+  - @tldraw/core@1.18.1
+
+## 1.24.0
+
+### Minor Changes
+
+- - Remove `mobx` and `mobx-react-lite` as dependencies. This is a breaking change for libraries that expect data to be observable in `@tldraw/core`.
+
+### Patch Changes
+
+- Updated dependencies
+  - @tldraw/core@1.18.0
+
+## 1.23.0
+
+### Minor Changes
+
+- - Improve middle mouse panning
+  - Fix bug with assets in VS Code plugin
+  - Improve performance of draw-style shapes
+  - Fix bug with creating assets
+  - Fix bug with text align in labels when outputting images
+  - Fix bug with middle mouse panning on Linux
+  - Fix bug with zoom shortcuts on number pad
+  - Fix bug with draw and erase direction when holding shift
+
+### Patch Changes
+
+- Updated dependencies
+  - @tldraw/core@1.17.0
+
+## 1.22.1
+
+### Patch Changes
+
+- Remove share by URL.
+
+## 1.22.0
+
+### Minor Changes
+
+- - Fix broken VS Code extension
+  - Add share by URL
+
+## 1.21.0
+
+### Minor Changes
+
+- - Improve text (and multiline text) in image exports
+  - Create European Portugese translation
+  - Create Swedish translation
+  - Use system default for theme default
+  - Update translation label for Chinese
+  - Fix bugs with flip command
+  - Fix bug with duplicate page command
+  - Improve dialogs
+  - Improve SVG pasting
+
+### Patch Changes
+
+- Updated dependencies
+  - @tldraw/core@1.16.0
+
 ## 1.20.0
 
 ### Minor Changes